Make defining .custom_jobs and .custom_perf optional
We define a default for those templates so that user can get started
with using Shared CI more easily. Customization is still possible by
overriding the template, in a dedicated file include in the child
pipeline, or in each .gitlab/jobs/<machine>.yml.

We adapted the docs, the examples, and fixed some things.