Run perf jobs if the perf allocation was given
Only run perf jobs for a given machine in allocation information was
provided.

By default the allocation information is empty, which will prevent the
jobs from running on this machine.

We split the config to create machine specific templates that allow to
apply the same behavior on other jobs the users may create for this
pipeline.

This is different than the `ON_MACHINE` variable since it only applies
to the performance pipeline and is not meant to be used globally: e.g.
pipeline jobs for a given machine cannot be turned off from the UI.